+++ pkg/apis/policy/v1alpha1/message.go
@@ -1,7 +1,7 @@
 package v1alpha1
 
 import (
-	"github.com/kyverno/kyverno-json/pkg/message"
+	"github.com/kyverno/kyverno-json/pkg/syntax/message"
 	"k8s.io/apimachinery/pkg/util/json"
 )
